refactor(components): Relocate shipping time input to shared components
