+++ js/src/components/free-listings/configure-product-listings/shipping-time-setup.js
@@ -8,7 +8,7 @@ import { __ } from '@wordpress/i18n';
  */
 import { useAdaptiveFormContext } from '~/components/adaptive-form';
 import Section from '~/components/section';
-import CountriesTimeInput from './countries-time-input';
+import CountriesTimeInput from '~/components/countries-time-input';
 
 /**
  * Form control to edit shipping time settings.
